Fixed a bug in highlighting
When highlighting text, if there was identical text (two matches in a document) the first would always be selected. This should now work. I fixed it by estimating the cursor position in javascript and then finding the closest match in the php code. There is never going to be a 1 to 1 match with position, but this seems to work.